+++ src/main/java/gift/api/member/controller/MemberController.java
@@ -1,7 +1,6 @@
 package gift.api.member.controller;
 
 import gift.api.member.dto.MemberRequest;
-import gift.api.member.service.MemberFacade;
 import gift.api.member.service.MemberService;
 import gift.global.utils.JwtUtil;
 import io.swagger.v3.oas.annotations.Operation;
@@ -9,12 +8,10 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -22,11 +19,9 @@ import org.springframework.web.bind.annotation.RestController;
 @Tag(name = "Member", description = "Member API")
 public class MemberController {
 
-    private final MemberFacade memberFacade;
     private final MemberService memberService;
 
-    public MemberController(MemberFacade memberFacade, MemberService memberService) {
-        this.memberFacade = memberFacade;
+    public MemberController(MemberService memberService) {
         this.memberService = memberService;
     }
 
@@ -43,13 +38,6 @@ public class MemberController {
     @Operation(summary = "로그인", description = "일반 로그인")
     public ResponseEntity<Void> login(@RequestBody MemberRequest memberRequest, @RequestHeader("Authorization") String token) {
         memberService.login(memberRequest, token.split(" ")[1]);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/oauth/kakao")
-    @Operation(summary = "카카오 로그인")
-    public ResponseEntity<Void> loginKakao(@RequestParam("code") String code) {
-        memberFacade.loginKakao(code);
         return ResponseEntity.ok().build();
     }
 }
